[BrowserKit] remove deprecated code
